fix: Resolve NaN loss issues in Bench2Drive training with comprehensive solution
## Problem
Training DiffusionDrive on Bench2Drive dataset resulted in NaN losses due to:
- Camera images not normalized (feeding [0,255] instead of [0,1])
- Different LiDAR normalization (log vs clip-and-divide)
- BEV semantic maps using float32 instead of integer labels
- Missing data validation and gradient explosion prevention

## Core Fixes

### Data Normalization
- **Camera**: Added /255.0 normalization in transfuser_features_b2d.py
- **LiDAR**: Aligned with NavSim approach using hist_max_per_pixel clipping
- **BEV Maps**: Changed to uint8 integer labels [0,6] in bev_semantic_utils.py

### Training Robustness
- Added autograd anomaly detection for debugging NaN sources
- Implemented gradient clipping (max_norm=1.0) in agent_lightning_module.py
- Added feature validation to catch NaN/Inf early in pipeline

## Infrastructure Improvements

### Cache Management
- Added NAVSIM_CACHE_ROOT environment variable in common.sh
- Organized cache paths under /workspace/navsim_workspace/cache
- Added safety warnings for Bench2Drive training in train.sh

### Debug Tools Organization
- Created navsim/planning/utils/data_validation/ for validation tools
- Created navsim/planning/utils/debugging/ for training debugging
- Added scripts/data_processing/ for cache regeneration tools
- Provided convenience wrapper script for all debug tools

## New Features
- Comprehensive data validation script with range checking
- Training health monitoring with real-time NaN detection
- Safe cache regeneration with backup capabilities
- Status checker to verify all fixes are applied
- Detailed documentation with step-by-step troubleshooting

## Testing
- All tools include validation and error handling
- Features validated against expected NavSim data ranges
- Gradient norms logged for monitoring training stability

This commit provides a complete solution to NaN loss issues with proper
tooling, documentation, and safety measures for robust training.

+++ navsim/common/bev_semantic_utils.py
@@ -299,23 +299,24 @@ def generate_simple_bev_semantic(
     if map_bev is not None:
         bev_map = map_bev.copy()
     else:
-        bev_map = np.zeros((bev_height, bev_width), dtype=np.float32)
+        bev_map = np.zeros((bev_height, bev_width), dtype=np.uint8)
         
         # Add road from trajectory only if no map data
         if trajectory is not None and len(trajectory) > 0:
             road_mask = generate_road_mask_from_trajectory(
                 trajectory, bev_height, bev_width, resolution
             )
-            bev_map[road_mask > 0] = 1.0  # Road class
+            bev_map[road_mask > 0] = 1  # Road class
     
     # Add vehicles from agents (always add dynamic objects)
     if agents is not None and agent_labels is not None:
         vehicle_mask = generate_vehicle_mask_from_agents(
             agents, agent_labels, bev_height, bev_width, resolution
         )
-        bev_map[vehicle_mask > 0] = 5.0  # Vehicle class
+        bev_map[vehicle_mask > 0] = 5  # Vehicle class
     
-    return bev_map
+    # Ensure integer type for class labels
+    return bev_map.astype(np.uint8)
 
 
 def project_semantic_to_bev(
